fix show fps bug

+++ Tsukuyomi/Tsukuyomi/D3DRenderer.cpp
@@ -515,10 +515,10 @@ void D3DRenderer::resetCameraTransform(Camera* cam)
 	m_camera.updateViewMatrix();
 }
 
-int D3DRenderer::computeFPS()
+float D3DRenderer::computeFPS()
 {
 	clock_t cur_clock = clock();
-	float passed_secs = std::max<float>((cur_clock - last_clock) / CLOCKS_PER_SEC, 0.001);
+	float passed_secs = std::max<float>((float)(cur_clock - last_clock) / CLOCKS_PER_SEC, 0.001);
 	last_clock = cur_clock;
 	return 1.0 / passed_secs;
 }
